fix: split up files in the salesforcedx-vscode-core/src/commands/util directory
