+++ packages/salesforcedx-vscode-core/src/commands/util/index.ts
@@ -5,9 +5,16 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+export { CommandletExecutor } from './commandletExecutor';
+export { CommandParams } from './commandParams';
+export { CompositePostconditionChecker } from './compositePostconditionChecker';
+export { CompositePreconditionChecker } from './compositePreconditionChecker';
 export { ConflictDetectionMessages } from './conflictDetectionMessages';
 export { createComponentCount } from './createComponentCount';
+export { DevUsernameChecker } from './devUsernameChecker';
+export { EmptyPreChecker } from './emptyPreChecker';
 export { EmptyPostChecker } from './emptyPostChecker';
+export { FlagParameter } from './flagParameter';
 export { formatException } from './formatException';
 export { LibraryPathsGatherer } from './libraryPathsGatherer';
 export { OverwriteComponentPrompt } from './overwriteComponentPrompt';
@@ -25,19 +32,9 @@ export {
   SelectFileName,
   SelectUsername
 } from './parameterGatherers';
-export {
-  SfdxWorkspaceChecker,
-  CompositePreconditionChecker,
-  DevUsernameChecker,
-  EmptyPreChecker
-} from './preconditionCheckers';
-export {
-  CommandletExecutor,
-  CommandParams,
-  FlagParameter,
-  SfdxCommandlet,
-  SfdxCommandletExecutor
-} from './sfdxCommandlet';
+export { SfdxCommandlet } from './sfdxCommandlet';
+export { SfdxCommandletExecutor } from './sfdxCommandletExecutor';
+export { SfdxWorkspaceChecker } from './sfdxWorkspaceChecker';
 export {
   SourcePathStrategy,
   PathStrategyFactory
