+++ packages/salesforcedx-vscode-core/src/commands/util/sfdxCommandlet.ts
@@ -4,129 +4,18 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+
 import {
-  CliCommandExecutor,
-  Command,
-  CommandExecution, ContinueResponse, Measurements, ParametersGatherer,
+  ParametersGatherer,
   PostconditionChecker,
-  PreconditionChecker, Properties,
-  TelemetryData
+  PreconditionChecker
 } from '@salesforce/salesforcedx-utils-vscode';
 import * as vscode from 'vscode';
 import { channelService } from '../../channels';
-import { notificationService, ProgressNotification } from '../../notifications';
+import { notificationService } from '../../notifications';
 import { sfdxCoreSettings } from '../../settings';
-import { taskViewService } from '../../statuses';
-import { telemetryService } from '../../telemetry';
-import { workspaceUtils } from '../../util';
+import { CommandletExecutor } from './commandletExecutor';
 import { EmptyPostChecker } from './emptyPostChecker';
-
-export interface FlagParameter<T> {
-  flag?: T;
-}
-
-export interface CommandParams {
-  readonly command: string;
-  // handle to localized user facing help text, with entries for diff flags
-  description: Record<string, string>;
-  logName: Record<string, string>; // metric key
-}
-
-export interface CommandletExecutor<T> {
-  execute(response: ContinueResponse<T>): void;
-  readonly onDidFinishExecution?: vscode.Event<[number, number]>;
-}
-
-export abstract class SfdxCommandletExecutor<T>
-  implements CommandletExecutor<T> {
-  protected showChannelOutput = true;
-  protected executionCwd = workspaceUtils.getRootWorkspacePath();
-  protected onDidFinishExecutionEventEmitter = new vscode.EventEmitter<
-    [number, number]
-  >();
-  public readonly onDidFinishExecution: vscode.Event<[number, number]> = this
-    .onDidFinishExecutionEventEmitter.event;
-
-  protected attachExecution(
-    execution: CommandExecution,
-    cancellationTokenSource: vscode.CancellationTokenSource,
-    cancellationToken: vscode.CancellationToken
-  ) {
-    channelService.streamCommandOutput(execution);
-
-    if (this.showChannelOutput) {
-      channelService.showChannelOutput();
-    }
-
-    notificationService.reportCommandExecutionStatus(
-      execution,
-      cancellationToken
-    );
-    ProgressNotification.show(execution, cancellationTokenSource);
-    taskViewService.addCommandExecution(execution, cancellationTokenSource);
-  }
-
-  public logMetric(
-    logName: string | undefined,
-    hrstart: [number, number],
-    properties?: Properties,
-    measurements?: Measurements
-  ) {
-    telemetryService.sendCommandEvent(
-      logName,
-      hrstart,
-      properties,
-      measurements
-    );
-  }
-
-  public execute(response: ContinueResponse<T>): void {
-    const startTime = process.hrtime();
-    const cancellationTokenSource = new vscode.CancellationTokenSource();
-    const cancellationToken = cancellationTokenSource.token;
-    const execution = new CliCommandExecutor(this.build(response.data), {
-      cwd: this.executionCwd,
-      env: { SFDX_JSON_TO_STDOUT: 'true' }
-    }).execute(cancellationToken);
-
-    let output = '';
-    execution.stdoutSubject.subscribe(realData => {
-      output += realData.toString();
-    });
-
-    execution.processExitSubject.subscribe(exitCode => {
-      const telemetryData = this.getTelemetryData(
-        exitCode === 0,
-        response,
-        output
-      );
-      let properties;
-      let measurements;
-      if (telemetryData) {
-        properties = telemetryData.properties;
-        measurements = telemetryData.measurements;
-      }
-      this.logMetric(
-        execution.command.logName,
-        startTime,
-        properties,
-        measurements
-      );
-      this.onDidFinishExecutionEventEmitter.fire(startTime);
-    });
-    this.attachExecution(execution, cancellationTokenSource, cancellationToken);
-  }
-
-  protected getTelemetryData(
-    success: boolean,
-    response: ContinueResponse<T>,
-    output: string
-  ): TelemetryData | undefined {
-    return;
-  }
-
-  public abstract build(data: T): Command;
-}
 
 export class SfdxCommandlet<T> {
   private readonly prechecker: PreconditionChecker;
