+++ packages/salesforcedx-vscode-core/test/jest/commands/templates/forceApexTriggerCreate.test.ts
@@ -19,13 +19,13 @@ import {
   SelectFileName,
   SelectOutputDir
 } from '../../../../src/commands/util/parameterGatherers';
-import { SfdxWorkspaceChecker } from '../../../../src/commands/util/preconditionCheckers';
 import * as commandlet from '../../../../src/commands/util/sfdxCommandlet';
+import { SfdxWorkspaceChecker } from '../../../../src/commands/util/sfdxWorkspaceChecker';
 
 jest.mock('../../../../src/commands/templates/executors/libraryForceApexTriggerCreateExecutor');
 jest.mock('../../../../src/commands/util/overwriteComponentPrompt');
 jest.mock('../../../../src/commands/util/parameterGatherers');
-jest.mock('../../../../src/commands/util/preconditionCheckers');
+jest.mock('../../../../src/commands/util/sfdxWorkspaceChecker');
 
 const selectFileNameMocked = jest.mocked(SelectFileName);
 const metadataTypeGathererMocked = jest.mocked(MetadataTypeGatherer);
@@ -55,6 +55,7 @@ describe('forceApexTriggerCreate Unit Tests.', () => {
         };
       });
   });
+
   it('Should be able to execute forceApexTriggerCreate.', async () => {
     await forceApexTriggerCreate();
     expect(selectFileNameMocked).toHaveBeenCalledWith(
